@@ -38,7 +38,7 @@ use Wikimedia\TestingAccessWrapper;
  */
 class DonationInterface_Adapter_GatewayAdapterTest extends DonationInterfaceTestCase {
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$this->setMwGlobals( [
