@@ -34,7 +34,7 @@ class DonationInterface_Adapter_GlobalCollect_RealTimeBankTransferIdealTest exte
 	 */
 	protected $bankPaymentProvider;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$config = TestingProviderConfiguration::createForProvider(
