@@ -36,7 +36,7 @@ class DonationInterface_Adapter_GlobalCollect_RecurringTest extends DonationInte
 		$this->testAdapterClass = TestingGlobalCollectAdapter::class;
 	}
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$this->setMwGlobals( [
