@@ -24,7 +24,7 @@
  * @group Yandex
  */
 class DonationInterface_Adapter_GlobalCollect_YandexTest extends DonationInterfaceTestCase {
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 
 		$this->setMwGlobals( [
