@@ -41,7 +41,7 @@ class AmountTest extends DonationInterfaceTestCase {
 	 */
 	protected $normalized;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 		$this->setMwGlobals( [
 			'wgDonationInterfacePriceFloor' => 1.50,
@@ -273,7 +273,9 @@ class AmountTest extends DonationInterfaceTestCase {
 		$this->assertEquals( '$100.59', $amount );
 
 		$amount = $validator->format( 100.59, 'USB', 'en_US' );
-		$this->assertEquals( 'USB100.59', $amount );
+		// Unknown currency code is just plunked next to amount
+		// with a non-breaking space in between.
+		$this->assertEquals( "USB\u{00A0}100.59", $amount );
 
 		$amount = $validator->format( 100.59, 'USD', 'en_CA' );
 		$this->assertEquals( 'US$100.59', $amount );
